requirements.txt
openpyxl
pandas
numpy
primer3
Biopython